Remove unnecessary tests and copyHandlerUtil function
