+++ azbfs/zz_generated_models.go
@@ -10,7 +10,6 @@ import (
 	"reflect"
 	"strconv"
 	"strings"
-	"time"
 )
 
 // concatenates a slice of const values with the specified separator between each item
@@ -406,19 +405,6 @@ type Path struct {
 	Owner       *string `json:"owner,omitempty"`
 	Group       *string `json:"group,omitempty"`
 	Permissions *string `json:"permissions,omitempty"`
-}
-
-func (p Path) LastModifiedTime() time.Time {
-	if p.LastModified == nil {
-		return time.Time{}
-	}
-
-	t, err := time.Parse(time.RFC1123, *p.LastModified)
-	if err != nil {
-		return time.Time{}
-	}
-
-	return t
 }
 
 func (p Path) ContentMD5() []byte {
